+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
  It checks if a the file or folder exists.
  If it's a path to a file, it make sure that it's in a format supported by REAPER.
  You can specify a specific extension or a list of extensions and it will raise if the choosen file does not match
- 
+ -  `Project.select_item()` method to select a specific item.
 ## [Unreleased] - 2020-12-30
 
 ### Fixed